oauth2_clientmanager: replace write-and-rename writes with atomic_write
The write-and-rename approach works but only if the file will always be
overwritten.  If we want to ensure that the session file isn't
accidentally overwritten if multiple instances are started, that check
needs to be stricter.  Rather than open code all of that, we can just
use the common atomicwrites module instead.