oauth2-clientd: fix stdout vs stderr when printing access token
If no socket or file is specified to update periodically, the tool will
exit and dump the access token to stdout.  The messages around it are
also dumped to stdout, which would make it a challenge to capture the
access token for use in a script.  This change prints the messages to
stderr instead so that a quiet run could be used to generate a single
access token easily.