oauth2-clientd: add suse-o365 provider and use it by default
This adds the ability to recursively inherit registration definitions
and uses that to add a 'suse-o365' provider that inherits from
the 'microsoft' provider and only extends with with tenant name,
client id, and (empty) client secret.

The suse-o365 provider is now the default.  If there is interest outside
of SUSE in using this tool, I can easily extend this to use a config
file for provider definitions instead of coding them in.  But for now,
it's easier to keep it self-contained.