oauth2-clientd: don't try to pick a random port and reuse it
In _setup_redirect_listener, we let the OS pick a random port, close it,
and try to reuse it.  This is a vestige of an older script that I had
initially based this on and it's unnecessary.  We can get the port
number from the server after it binds but before we activate it and
just let the OS pick the random port then.