oauth2-clientd: restructure package
As I want to use package resource files, it doesn't make a ton of sense
to keep the script separate from the package.  This renames the package
to oauth2_clientd and puts the clientmanager into a new module called
sessionmanager.  It pushes most of the functionality of the script into
a cli module and calls main from the script.