oath2-clientd: enable google registration support
Now that we've added token support for the Authorization Code Grant,
and allowed us to retain url authentication support, enable Google
as a registration method, and specify the requirements for each of
the other registrations.

So far google is our first token Authorization Code Grant registrar.

I've tested this with my gmail account using mbsync.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>