sessionmanager: break out helpers into new module
There are a few helpers that are either classmethods or staticmethods
that were there for organizational purposes.  As I prepare to break up
the sessionmanager module into a bunch of more easily tested parts,
these helpers will be shared among the modules and should be broken out
into a new module themselves.