cli: divide sleep time before refreshing token into threshold-long periods
When oauth2_clientd runs on a machine that gets suspended oauth2_clientd
does not update the token in a timely manner after the machine gets woken
up. Tokens stay valid for a period of time (usually an hour and 15
minutes). If the machine was suspended shortly after the token had been
renewed oauth2_clientd would wait almost all of the validity period to
renew the token again.

One possible solution is to divide the sleep time for which oauth2_clientd
blocks before refreshing the token into shorter periods. The threshold
parameter seems a natural choice since it has the advantage of not
introducing another time-related parameter.

Signed-off-by: Jiri Wiesner <jwiesner@suse.de>