cli: use CLOCK_BOOTTIME to handle waiting for refresh timeout (#7)
All of the mechanisms that Python supports natively to handle sleeping
use CLOCK_MONOTONIC either explicitly or implicitly.  This has the
advantage of the clock never jumping forward or backward, but the
disadvantage (at least for this purpose) of stopping while the system is
suspended.  This means that when oauth2-clientd is used on a system that
suspends, if the system is suspended for more than the configured
threshold number of seconds (cumulatively), we'll miss the refresh
expiration and the user will not be notified until the original sleep
has elapsed.

The solution should be to use the clock_nanosleep system call with the
CLOCK_BOOTTIME clock, which also never jumps forward or backward but is
active while the system is suspended.  Unfortunately, Python doesn't
offer a native interface to clock_nanosleep so we need to work around it
using clock_gettime with CLOCK_BOOTTIME and a short timeout in a loop to
approximate it.

Fixes: #7